--- proto/optional_output_wrapper/BUILD
@@ -1,7 +1,0 @@
-load("//rust:rust.bzl", "rust_binary")
-
-rust_binary(
-    name = "optional_output_wrapper",
-    srcs = ["bin.rs"],
-    visibility = ["//visibility:public"],
-)
